css improvements for token action hud
