+++ module.json
@@ -48,7 +48,8 @@
   ],
   "esmodules": [
     "scripts/character-sheet.js",
-    "scripts/custom-buttons.js"
+    "scripts/custom-buttons.js",
+    "scripts/tlm-token-action-hud.js"
   ],
   "url": "https://github.com/snaakn/the-longest-mile",
   "download": "https://github.com/snaakn/the-longest-mile/releases/download/v_VERSION/module.zip",
