+++ scripts/character-sheet.js
@@ -1,18 +1,6 @@
-import { MODULE_NAME } from "./consts.js";
+// import { MODULE_NAME } from "./consts.js";
+import { injectCSS } from "./utils.js";
 
 Hooks.once("init", async () => {
   injectCSS("character-sheet");
 });
-
-function injectCSS(filename) {
-  const head = document.getElementsByTagName("head")[0];
-  const mainCss = document.createElement("link");
-  mainCss.setAttribute("rel", "stylesheet");
-  mainCss.setAttribute("type", "text/css");
-  mainCss.setAttribute(
-    "href",
-    "modules/" + MODULE_NAME + "/styles/" + filename + ".css"
-  );
-  mainCss.setAttribute("media", "all");
-  head.insertBefore(mainCss, head.lastChild);
-}
