+++ styles/tlm-tah.css
@@ -1,0 +1,27 @@
+.tah-actions {
+    display: inline-flex !important;
+    width: 100% !important;
+}
+
+.tah-action {
+    max-width: 150px;
+}
+
+/* This just looks nicer*/
+.tah-action button {
+    padding: 1px 2px;
+    height: fit-content;
+}
+
+/* Setting tah-actions display to inline-flex fixes the issue of too small grid size being calculated. (e.g. Draw Dagger)
+   However, the status effects should be displayed in an even grid, so the buttons need the same size. */
+.tah-action button.toggle {
+    min-width: 130px;
+    max-width: 130px;
+} 
+
+/* The +/- buttons in spells were so big and distracting - this makes them smaller */
+.tah-action.shrink {
+    min-width: fit-content !important;
+    width: fit-content;
+}
